- Added ability to remove XP from job and sink level if needed  - Removed PricefactoryJobLevel
