+++ build.gradle
@@ -47,7 +47,7 @@ allprojects {
     }
 
     dependencies {
-        compileOnly 'com.willfp:eco:6.51.1'
+        compileOnly 'com.willfp:eco:6.51.3'
         implementation 'com.willfp:libreforge:3.129.2'
         implementation 'com.willfp:ecomponent:1.3.0'
         implementation 'org.joml:joml:1.10.4'
