+++ eco-core/core-plugin/src/main/kotlin/com/willfp/ecojobs/jobs/Jobs.kt
@@ -4,9 +4,7 @@ import com.google.common.collect.BiMap
 import com.google.common.collect.HashBiMap
 import com.google.common.collect.ImmutableList
 import com.willfp.eco.core.config.updating.ConfigUpdater
-import com.willfp.eco.core.price.Prices
 import com.willfp.ecojobs.EcoJobsPlugin
-import com.willfp.ecojobs.PriceFactoryJobLevel
 import com.willfp.ecojobs.api.getJobLevel
 import org.bukkit.OfflinePlayer
 
@@ -48,10 +46,6 @@ object Jobs {
 
         for ((id, jobConfig) in plugin.fetchConfigs("jobs")) {
             addNewJob(Job(id, jobConfig, plugin))
-        }
-
-        values().forEach {
-            Prices.registerPriceFactory(PriceFactoryJobLevel(it))
         }
     }
